chore: do code separation refactor
